[SPARK-24748][SS][FOLLOWUP] Switch custom metrics to Unstable APIs
- Mark custom metrics related APIs as unstable
- Move streaming related interfaces to streaming package
